fix: can't modify is not admin

+++ web/src/components/companies/CompaniesTable.js
@@ -1,8 +1,13 @@
 import MaterialTable from "material-table";
 import React, { useEffect } from "react";
 import { useDispatch, useSelector } from "react-redux";
-import IconsTable from './../../common/icons/IconsTable';
-import { createCompanyAction, deleteCompanyAction, getCompaniesAction, updateCompanyAction } from './../../redux/actions/companies.action';
+import IconsTable from "./../../common/icons/IconsTable";
+import {
+  createCompanyAction,
+  deleteCompanyAction,
+  getCompaniesAction,
+  updateCompanyAction
+} from "./../../redux/actions/companies.action";
 
 const columns = [
   { title: "Name", field: "name" },
@@ -14,11 +19,11 @@ const columns = [
 function CompaniesTable() {
   const [empData, setEmpData] = React.useState([]);
   const companies = useSelector((state) => state.company.companies);
-  console.log(empData);
+  const admin = useSelector((state) => state.user.admin);
   const dispatch = useDispatch();
 
   const loadCompanies = () => {
-    dispatch(getCompaniesAction())
+    dispatch(getCompaniesAction());
   };
 
   useEffect(() => {
@@ -26,49 +31,53 @@ function CompaniesTable() {
   }, []);
 
   useEffect(() => {
-    const editable = companies.map(o => ({ ...o }));
+    const editable = companies.map((o) => ({ ...o }));
     setEmpData(editable);
   }, [companies]);
-  
 
   return (
     <>
-    <MaterialTable
-          title="Companies list"
-          data={empData}
-          columns={columns}
-          options={{
-            search: true,
-            paging: true,
-            filtering: true,
-            exportButton: true
-          }}
-          editable={{
-            onRowAdd: (newData) =>
-              new Promise((resolve, reject) => {
-                setTimeout(() => {
+      <MaterialTable
+        title="Companies list"
+        data={empData}
+        columns={columns}
+        options={{
+          search: true,
+          paging: true,
+          filtering: true,
+          exportButton: true,
+        }}
+        editable={{
+          onRowAdd: (newData) =>
+            new Promise((resolve, reject) => {
+              setTimeout(() => {
+                if (admin) {
                   dispatch(createCompanyAction(newData));
-                  resolve();
-                }, 1000);
-              }),
-            onRowUpdate: (newData, oldData) =>
-              new Promise((resolve, reject) => {
-                setTimeout(() => {
+                }
+                resolve();
+              }, 1000);
+            }),
+          onRowUpdate: (newData, oldData) =>
+            new Promise((resolve, reject) => {
+              setTimeout(() => {
+                if (admin) {
                   dispatch(updateCompanyAction(newData));
-                  resolve();
-                }, 1000);
-              }),
-            onRowDelete: (oldData) =>
-              new Promise((resolve, reject) => {
-                setTimeout(() => {
+                }
+                resolve();
+              }, 1000);
+            }),
+          onRowDelete: (oldData) =>
+            new Promise((resolve, reject) => {
+              setTimeout(() => {
+                if (admin) {
                   dispatch(deleteCompanyAction(oldData.id));
-                  resolve();
-                }, 1000);
-              }),
-          }}
-          icons={IconsTable()}
-        />
-        
+                }
+                resolve();
+              }, 1000);
+            }),
+        }}
+        icons={IconsTable()}
+      />
     </>
   );
 }
